feat : 홈 추천 손해사정사 조회 훅 추가
* 손해사정사 목록 스키마·조회 함수를 customer/_shared로 승격(dashboard·adjusters 공유)
* 평점순 첫 페이지를 반환하는 useRecommendedAdjusters 조회 훅 추가
* adjuster 추천 전용 쿼리키 추가(목록 무한 조회와 캐시 분리)
* 기존 adjusters 세그먼트 소비처는 재노출로 무중단 유지
